requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
diffusers==0.35.1
transformers==4.55.2
python-multipart==0.0.20
python-dotenv==1.1.1
torch==2.5.1
Pillow==11.0.0
numpy==2.1.2
huggingface_hub==0.34.4
torchmetrics==1.6.0
